pyproject.toml
[tool.poetry]
name = "gemtel"
version = "0.1.0"
description = ""
authors = ["OverflowCat <overflowcat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gemini-webapi = "^1.2.0"
aiogram = "^3.6.0"
aiohttp-socks = "^0.8.4"
magic-filter = "^1.0.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
